+++ project/tests/GUI/MasterGUI/test_project_overview.py
@@ -1,18 +1,18 @@
 import unittest
 from unittest.mock import Mock
-from GUI.GUIFactory import SlaveOverview
+from GUI.MasterGUI.ProjectOverview import ProjectOverview
 from Domain.Master import Master
 from Domain.SlaveConnection import SlaveConnection
 from Domain.Presentation import Presentation
-from Domain.PathConstants import PathConstants
 
 
-class SlaveOverviewTest(unittest.TestCase):
+class ProjectOverviewTest(unittest.TestCase):
 
     def setUp(self):
-        self.slave_overview = SlaveOverview()
+        self.slave_overview = ProjectOverview()
         self.slave_connection_mock = SlaveConnection(Mock(Master))
-        presentation = ["a.jpg", "b.jpg", "g.mp4", "test_text.txt", "test_text2.txt"]
+        presentation = Presentation()
+        presentation.set_files(["a.jpg", "b.jpg", "g.mp4", "test_text.txt", "test_text2.txt"])
         self.slave_connection_mock.set_presentation(presentation)
 
     def test_update_slave_to_overview(self):
